backend/requirements.txt
Flask==3.0.3
Flask-Sock==0.7.0
sounddevice==0.4.6
scipy==1.13.1
librosa==0.10.2
numpy==1.26.4
google-generativeai==0.6.0
gTTS==2.5.1
pydub==0.25.1
transformers==4.41.2
scikit-learn==1.5.0

git+https://github.com/openai/whisper.git